tests(.): Split tests into different directories
