+++ backend/test/integration/thesis.unauthorized.integration.test.js
@@ -1,9 +1,9 @@
 require('jest');
 // [i] This line setup the test database + load the environment variables. DON'T (RE)MOVE IT
-const { resetTestDatabase } = require('./integration_config');
+const { resetTestDatabase } = require('../integration_config');
 
-const { app } = require("../app");
-const utils = require("./utils");
+const { app } = require("../../app");
+const utils = require("../utils");
 
 beforeEach(() => {
   // Be sure that we are using a full clean database before each test
